refactor ISR & IRQ code
